add test for "disabling parameter/retvals" for copydoc, copydetails...
If copydoc/copydetails is used in the body of a method, the
documentation blocks for parameters/return values are skipped.
Also the automatic documentation is ignored.
There is one exception, however:
The strings "object of type" are parsed and used in order to
specify the parameter's/return value's type.

See the static "test" method in classA and grid.rect.rectgrid
for details. The definition of the type for its first parameter
auto_param is obtained from the 'conf' file.